+++ all-is-cubes-desktop/src/terminal.rs
@@ -24,7 +24,7 @@ use tui::widgets::{Borders, Paragraph};
 use tui::{backend::CrosstermBackend, Terminal};
 
 use all_is_cubes::apps::{AllIsCubesAppState, Key, StandardCameras};
-use all_is_cubes::camera::{Camera, Viewport};
+use all_is_cubes::camera::{Camera, ToneMappingOperator, Viewport};
 use all_is_cubes::cgmath::Vector2;
 use all_is_cubes::cgmath::{ElementWise as _, Point2};
 use all_is_cubes::inv::Slot;
@@ -174,11 +174,12 @@ impl TerminalMain {
                         scene,
                     }) = render_thread_in.recv()
                     {
-                        let (image, info) =
-                            scene.get().trace_scene_to_image::<ColorCharacterBuf, _, _>(
-                                &camera,
-                                TextAndColor::from,
-                            );
+                        let tmo = &camera.options().tone_mapping;
+                        let (image, info) = scene
+                            .get()
+                            .trace_scene_to_image::<ColorCharacterBuf, _, _>(&camera, |b| {
+                                b.output(tmo)
+                            });
                         // Ignore send errors as they just mean we're shutting down or died elsewhere
                         let _ = render_thread_out.send(FrameOutput {
                             viewport: camera.viewport(),
@@ -828,6 +829,20 @@ struct ColorCharacterBuf {
     override_color: bool,
 }
 
+impl ColorCharacterBuf {
+    fn output(self, tone_map: &ToneMappingOperator) -> TextAndColor {
+        // TODO: override_color should be less clunky
+        if self.override_color {
+            (self.text.into(), None)
+        } else {
+            (
+                self.text.into(),
+                Some(Rgba::from(self.color).map_rgb(|rgb| tone_map.apply(rgb))),
+            )
+        }
+    }
+}
+
 impl PixelBuf for ColorCharacterBuf {
     type BlockData = <CharacterBuf as PixelBuf>::BlockData;
 
@@ -850,18 +865,6 @@ impl PixelBuf for ColorCharacterBuf {
         self.text
             .add(Rgba::TRANSPARENT, &CharacterRtData(Cow::Borrowed(" ")));
         self.override_color = true;
-    }
-}
-
-impl From<ColorCharacterBuf> for TextAndColor {
-    #[inline]
-    fn from(buf: ColorCharacterBuf) -> (String, Option<Rgba>) {
-        // TODO: override_color should be less clunky
-        if buf.override_color {
-            (buf.text.into(), None)
-        } else {
-            (buf.text.into(), Some(buf.color.into()))
-        }
     }
 }