+++ all-is-cubes/src/lum/glrender.rs
@@ -260,8 +260,13 @@ where
             .pipeline(
                 &self.back_buffer,
                 // TODO: port skybox cube map code
-                &PipelineState::default()
-                    .set_clear_color(world_output.data.sky_color.with_alpha_one().to_srgb_float()),
+                &PipelineState::default().set_clear_color(
+                    graphics_options
+                        .tone_mapping
+                        .apply(world_output.data.sky_color)
+                        .with_alpha_one()
+                        .to_srgb_float(),
+                ),
                 |pipeline, mut shading_gate| {
                     let world_output_bound = world_output.bind(&pipeline)?;
                     // Space