+++ all-is-cubes/src/lum/raytrace_to_texture.rs
@@ -103,14 +103,16 @@ where
                 .collect();
 
             let start_time = Instant::now();
+            let tmo = &camera.options().tone_mapping;
             let trace = |point: Point| {
                 let (color_buf, _info) =
                     tracer.trace_ray::<ColorBuf>(camera.project_ndc_into_world(Point2::new(
                         render_viewport.normalize_fb_x(point.x as usize),
                         render_viewport.normalize_fb_y(point.y as usize),
                     )));
-                // TODO: simplify this via `impl From<ColorBuf> for Rgb888`?
-                let [r, g, b, _a] = Rgba::from(color_buf).to_srgb_32bit();
+                let [r, g, b, _a] = Rgba::from(color_buf)
+                    .map_rgb(|rgb| tmo.apply(rgb))
+                    .to_srgb_32bit();
                 let color = Rgb888::new(r, g, b);
                 Pixel(point, color)
             };